modefied

+++ security/src/main/java/com/alibou/security/personnel/PersonnelRepository.java
@@ -2,5 +2,5 @@ package com.alibou.security.personnel;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonnelRepository extends JpaRepository<Personnel, Long> {
+public interface PersonnelRepository extends JpaRepository<Personnel, Integer> {
 }
